requirements.txt
uvicorn
aiohttp
aiosqlite
fastapi
nicegui
requests
discord_webhook
pyinstaller
slowapi
cryptography
notify-py
rich
pycryptodome